+++ syskit-ut/SpinSectionSuite.cpp
@@ -52,7 +52,7 @@ void SpinSectionSuite::testCtor01()
     // Create threads which will exercise some SpinSection code.
     Semaphore sem(0U);
     Thread** threadVec = new Thread*[16];
-    long i;
+    int i;
     for (i = 0; i < 16; ++i)
     {
         Thread* t = new Thread(entry00, &sem);
@@ -91,7 +91,7 @@ void* SpinSectionSuite::entry00(void* arg)
     sem->decrement(Semaphore::ETERNITY);
 
     bool ok = true;
-    for (unsigned long i = 0; i < 128; ++i)
+    for (unsigned int i = 0; i < 128; ++i)
     {
 
         if (ss_.tryLock())
@@ -108,7 +108,7 @@ void* SpinSectionSuite::entry00(void* arg)
         }
 
         // Update shared resource.
-        for (unsigned long j = 0; j < 64; ++j)
+        for (unsigned int j = 0; j < 64; ++j)
         {
             u64_ >>= 1;
             Thread::yield();
